internal/meta/layer: linearize WAL operations
This makes all WAL operations (including WALCheck) do writes to a row
based on which key they're checking the WAL of. This forces all WAL
operations on a single key to a linearizable history.

This fixes a race condition where the garbage collector incorrectly
sees that a WAL entry is missing from the database and deletes data that
was in the process of being written.

The garbage collector now commits more write transactions when it does
cleanup operations, but otherwise, the performance should be identical.

Fixes #7